Contoso done. Was problems with migration. I had to delete it all, clear history and remake it.

+++ ContosoUniversity/Controllers/CoursesController.cs
@@ -21,9 +21,9 @@ namespace ContosoUniversity.Controllers
 
         // GET: Courses
         public async Task<IActionResult> Index()
-        {
+        {   // returns Course entities (courses instead of schoolContext
             var courses = _context.Courses
-                .Include(c => c.Department);
+            .Include(c => c.Department);
             //.AsNoTracking();
             return View(await courses.ToListAsync());
         }
@@ -31,7 +31,7 @@ namespace ContosoUniversity.Controllers
         // GET: Courses/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Courses == null)
             {
                 return NotFound();
             }
@@ -47,42 +47,16 @@ namespace ContosoUniversity.Controllers
 
             return View(course);
         }
-
-        // GET: Courses/Create
-        //public IActionResult Create()
-        //{
-        //    ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "DepartmentID");
-        //    return View();
-        //}
-
         public IActionResult Create()
         {
             PopulateDepartmentsDropDownList();
             return View();
         }
-
-        // POST: Courses/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        //[HttpPost]
-        //[ValidateAntiForgeryToken]
-        //public async Task<IActionResult> Create([Bind("CourseID,Title,Credits,DepartmentID")] Course course)
-        //{
-        //    if (ModelState.IsValid)
-        //    {
-        //        _context.Add(course);
-        //        await _context.SaveChangesAsync();
-        //        return RedirectToAction(nameof(Index));
-        //    }
-        //    ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "DepartmentID", course.DepartmentID);
-        //    return View(course);
-        //}
-        
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("CourseID,Credits,DepartmentID,Title")] Course course)
         {
-            if (ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
                 _context.Add(course);
                 await _context.SaveChangesAsync();
@@ -91,23 +65,6 @@ namespace ContosoUniversity.Controllers
             PopulateDepartmentsDropDownList(course.DepartmentID);
             return View(course);
         }
-
-        // GET: Courses/Edit/5
-        //public async Task<IActionResult> Edit(int? id)
-        //{
-        //    if (id == null)
-        //    {
-        //        return NotFound();
-        //    }
-
-        //    var course = await _context.Courses.FindAsync(id);
-        //    if (course == null)
-        //    {
-        //        return NotFound();
-        //    }
-        //    ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "DepartmentID", course.DepartmentID);
-        //    return View(course);
-        //}
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -116,50 +73,20 @@ namespace ContosoUniversity.Controllers
             }
 
             var course = await _context.Courses
-                .AsNoTracking()
-                .FirstOrDefaultAsync(m => m.CourseID == id);
+                    //.AsNoTracking()
+                    .FirstOrDefaultAsync(m => m.CourseID == id);
+
             if (course == null)
             {
                 return NotFound();
             }
+
             PopulateDepartmentsDropDownList(course.DepartmentID);
             return View(course);
-        }
-        // POST: Courses/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        //[HttpPost]
-        //[ValidateAntiForgeryToken]
-        //public async Task<IActionResult> Edit(int id, [Bind("CourseID,Title,Credits,DepartmentID")] Course course)
-        //{
-        //    if (id != course.CourseID)
-        //    {
-        //        return NotFound();
-        //    }
 
-        //    if (ModelState.IsValid)
-        //    {
-        //        try
-        //        {
-        //            _context.Update(course);
-        //            await _context.SaveChangesAsync();
-        //        }
-        //        catch (DbUpdateConcurrencyException)
-        //        {
-        //            if (!CourseExists(course.CourseID))
-        //            {
-        //                return NotFound();
-        //            }
-        //            else
-        //            {
-        //                throw;
-        //            }
-        //        }
-        //        return RedirectToAction(nameof(Index));
-        //    }
-        //    ViewData["DepartmentID"] = new SelectList(_context.Departments, "DepartmentID", "DepartmentID", course.DepartmentID);
-        //    return View(course);
-        //}
+
+        }
+
         [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> EditPost(int? id)
@@ -169,6 +96,7 @@ namespace ContosoUniversity.Controllers
                 return NotFound();
             }
 
+
             var courseToUpdate = await _context.Courses
                 .FirstOrDefaultAsync(c => c.CourseID == id);
 
@@ -176,34 +104,37 @@ namespace ContosoUniversity.Controllers
                 "",
                 c => c.Credits, c => c.DepartmentID, c => c.Title))
             {
-                try
-                {
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateException /* ex */)
-                {
-                    //Log the error (uncomment ex variable name and write a log.)
-                    ModelState.AddModelError("", "Unable to save changes. " +
-                        "Try again, and if the problem persists, " +
-                        "see your system administrator.");
-                }
-                return RedirectToAction(nameof(Index));
+                PopulateDepartmentsDropDownList(courseToUpdate.DepartmentID);
+                return View(courseToUpdate);
             }
-            PopulateDepartmentsDropDownList(courseToUpdate.DepartmentID);
-            return View(courseToUpdate);
-        }
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateException /* ex */)
+            {
+                //Log the error (uncomment ex variable name and write a log.)
+                ModelState.AddModelError("", "Unable to save changes. " +
+                    "Try again, and if the problem persists, " +
+                    "see your system administrator.");
+            }
+            return RedirectToAction(nameof(Index));
 
-        private void PopulateDepartmentsDropDownList(object selectedDepartment = null)
+        }
+        private void PopulateDepartmentsDropDownList(object? selectedDepartment = null)
         {
             var departmentsQuery = from d in _context.Departments
                                    orderby d.Name
                                    select d;
             ViewBag.DepartmentID = new SelectList(departmentsQuery.AsNoTracking(), "DepartmentID", "Name", selectedDepartment);
         }
+
+
+
         // GET: Courses/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Courses == null)
             {
                 return NotFound();
             }
@@ -225,15 +156,24 @@ namespace ContosoUniversity.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Courses == null)
+            {
+                return Problem("Entity set 'SchoolContext.Courses'  is null.");
+            }
             var course = await _context.Courses.FindAsync(id);
-            _context.Courses.Remove(course);
+            if (course != null)
+            {
+                _context.Courses.Remove(course);
+            }
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
+
         private bool CourseExists(int id)
         {
-            return _context.Courses.Any(e => e.CourseID == id);
+            return (_context.Courses?.Any(e => e.CourseID == id)).GetValueOrDefault();
         }
     }
 }
